Loop over repository IDs
This makes it trivially easy to add more repositories in the future.

Note the /proc/self/fd/2 workaround for /dev/stderr; hopefully this will
be fixed on builds.sr.ht in the future. /dev/stderr is present on the
two other images I tried, archlinux and openbsd. See possible relevant
StackOverflow answer[0].

[0]: https://unix.stackexchange.com/a/38580